+++ src/decisions/admin/edit.controller.spec.ts
@@ -368,6 +368,117 @@ describe('EditController', () => {
       });
     });
 
+    describe('when the `action` is "submit"', () => {
+      it('saves the given dataset as draft and redirects to the submission confirmation page', async () => {
+        const editDto: EditDto = {
+          routes: ['Example route'],
+          countries: [['Brazil']],
+          yeses: [['1']],
+          noes: [['5']],
+          yesAfterComps: [['4']],
+          noAfterComps: [['7']],
+          action: 'submit',
+        };
+
+        const decisionRoutes: DecisionRoute[] = [
+          {
+            name: 'Example route',
+            countries: [
+              {
+                code: 'BR',
+                decisions: {
+                  yes: 1,
+                  no: 5,
+                  yesAfterComp: 5,
+                  noAfterComp: 7,
+                },
+              },
+            ],
+          },
+        ];
+
+        const profession = professionFactory.build({
+          id: 'example-profession-id',
+        });
+
+        const organisation = organisationFactory.build({
+          id: 'example-organisation-id',
+        });
+
+        const user = userFactory.build();
+
+        professionsService.findWithVersions.mockResolvedValue(profession);
+        organisationsService.find.mockResolvedValue(organisation);
+        decisionDatasetsService.find.mockResolvedValue(undefined);
+
+        const checkCanChangeDatasetSpy = jest
+          .spyOn(checkCanChangeDatasetModule, 'checkCanChangeDataset')
+          .mockImplementation();
+
+        const getActingUserSpy = jest
+          .spyOn(getActingUserModule, 'getActingUser')
+          .mockReturnValue(user);
+
+        const parseEditDtoDecisionRoutesSpy = jest
+          .spyOn(parseEditDtoDecisionRoutesModule, 'parseEditDtoDecisionRoutes')
+          .mockReturnValue(decisionRoutes);
+        const modifyDecisionRoutesSpy = jest.spyOn(
+          modifyDecisionRoutesModule,
+          'modifyDecisionRoutes',
+        );
+
+        const request = createDefaultMockRequest();
+        const response = createMock<Response>();
+        const flashMock = flashMessage as jest.Mock;
+        flashMock.mockImplementation(() => 'STUB_FLASH_MESSAGE');
+
+        await controller.update(
+          'example-profession-id',
+          'example-organisation-id',
+          2016,
+          editDto,
+          request,
+          response,
+        );
+
+        expect(response.redirect).toHaveBeenCalledWith(
+          '/admin/decisions/example-profession-id/example-organisation-id/2016/submit?fromEdit=true',
+        );
+
+        expect(professionsService.findWithVersions).toHaveBeenCalledWith(
+          'example-profession-id',
+        );
+        expect(organisationsService.find).toHaveBeenCalledWith(
+          'example-organisation-id',
+        );
+        expect(decisionDatasetsService.find).toHaveBeenCalledWith(
+          'example-profession-id',
+          'example-organisation-id',
+          2016,
+        );
+
+        expect(decisionDatasetsService.save).toHaveBeenCalledWith({
+          profession,
+          organisation,
+          year: 2016,
+          user,
+          status: DecisionDatasetStatus.Draft,
+          routes: decisionRoutes,
+        } as DecisionDataset);
+
+        expect(checkCanChangeDatasetSpy).toHaveBeenCalledWith(
+          request,
+          profession,
+          organisation,
+          2016,
+          false,
+        );
+        expect(getActingUserSpy).toHaveBeenCalledWith(request);
+        expect(parseEditDtoDecisionRoutesSpy).toHaveBeenCalledWith(editDto);
+        expect(modifyDecisionRoutesSpy).not.toHaveBeenCalled();
+      });
+    });
+
     describe('when the `action` is "save"', () => {
       it('saves the given dataset with the draft status', async () => {
         const editDto: EditDto = {