+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add confirmation page when publishing decision data from the editing page
 
+### Changed
+
+- Allow organisation users to submit decision data directly from the editing page
+
 ## [release-019] - 2022-05-04
 
 ### Changed